Added UnsafeBuffer utility class to simplify the management of unsafe buffers
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@205 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ SDRSharp/MainForm.cs
@@ -187,7 +187,7 @@ namespace SDRSharp
             while (_audioControl.IsPlaying)
             {
                 // http://www.designnews.com/author.asp?section_id=1419&doc_id=236273&piddl_msgid=522392
-                var fftGain = (float) (10.0 * Math.Log10(_fftBins / 2));
+                var fftGain = (float) (10.0 * Math.Log10((double) _fftBins / 2));
                 var compensation = 24.0f - fftGain;
 
                 while (_audioControl.IsPlaying)
